scheduling offsites

+++ solution_1/workday_time.rb
@@ -5,7 +5,7 @@ module WorkdayTime
     def time_string_to_numeric(time_string)
       raise "Not a valid time string" unless time_string.include?(':')
       hours = time_string.split(':')[0].to_i
-      minutes = time_string.split(':')[1].to_i / 60
+      minutes = time_string.split(':')[1].to_f / 60
       hours + minutes
     end
 
@@ -18,7 +18,6 @@ module WorkdayTime
 
     def add_time_length_to_start_time(time_length, start_time = START_OF_DAY)
       raise "time_length should be a number" unless time_length.is_a? Numeric
-
       end_time = time_string_to_numeric(start_time) + time_length
       numeric_time_to_time_string(end_time)
     end
